Added navigation functionality
