+++ App.tsx
@@ -2,92 +2,25 @@
 
 import React from 'react';
 import { StyleSheet, Text, View, Image, TouchableOpacity } from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';
+import {createNativeStackNavigator} from "@react-navigation/native-stack";
+import HomeScreen from "./screens/HomeScreen";
+import SportScreen from "./screens/SportScreen";
+import ElectricScreen from "./screens/ElectricScreen";
+import MovingScreen from "./screens/MovingScreen";
+import FamilyScreen from "./screens/FamilyScreen";
 
 export default function App() {
+  const Stack = createNativeStackNavigator();
   return (
-    <View style={styles.container}>
-      <Text style={styles.header}>RentRide</Text>
-      <View style={styles.menu}>
-        <Text>All</Text>
-        <Text>Sport</Text>
-        <Text>Electric</Text>
-        <Text>Moving</Text>
-        <Text>Family</Text>
-      </View>
-      <Car
-        imageSrc={require('./assets/mercedes.jpg')}
-        name="Mercedes C63"
-        fuel="Gas"
-        transmission="Manual"
-        seats={5}
-        price="55.99$"
-      />
-      <Car
-        imageSrc={require('./assets/vwpolo.jpg')}
-        name="VW Polo Aut."
-        fuel="Gas"
-        transmission="Automatic"
-        seats={5}
-        price="45.99$"
-      />
-    </View>
+  <NavigationContainer>
+    <Stack.Navigator initialRouteName='Home'>
+      <Stack.Screen name="Home" component={HomeScreen}/>
+      <Stack.Screen name="Sport" component={SportScreen}/>
+      <Stack.Screen name="Electric" component={ElectricScreen}/>
+      <Stack.Screen name="Moving" component={MovingScreen}/>
+      <Stack.Screen name="Family" component={FamilyScreen}/>
+    </Stack.Navigator>
+  </NavigationContainer>
   );
 }
-
-const Car: React.FC<{
-  imageSrc: any;
-  name: string;
-  fuel: string;
-  transmission: string;
-  seats: number;
-  price: string;
-}> = ({ imageSrc, name, fuel, transmission, seats, price }) => {
-  return (
-    <View style={styles.carContainer}>
-      <Image source={imageSrc} style={styles.carImage} />
-      <Text>{name}</Text>
-      <Text>Fuel: {fuel}</Text>
-      <Text>Transmission: {transmission}</Text>
-      <Text>Seats: {seats}</Text>
-      <Text>{price} / day</Text>
-      <TouchableOpacity style={styles.moreInfoButton}>
-        <Text>More info</Text>
-      </TouchableOpacity>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    paddingHorizontal: 20,
-  },
-  header: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  menu: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 20,
-  },
-  carContainer: {
-    marginBottom: 20,
-  },
-  carImage: {
-    width: '100%',
-    height: 200,
-  },
-  moreInfoButton: {
-    marginTop: 10,
-    padding: 10,
-    backgroundColor: '#f1f1f1',
-    borderRadius: 5,
-    alignItems: 'center',
-  },
-});
-
-
-
