+++ package.json
@@ -10,11 +10,15 @@
   },
   "dependencies": {
     "@expo/webpack-config": "^19.0.0",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/native-stack": "^6.9.17",
     "expo": "~49.0.10",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.72.4",
+    "react-native-safe-area-context": "^4.6.3",
+    "react-native-screens": "~3.22.0",
     "react-native-web": "~0.19.6"
   },
   "devDependencies": {
